ci.ocp: Sort images according to git
The quay.io registry returns the tags sorted alphabetically and doesn't
seem to provide a way to sort it by age. Let's use not really efficient
but functional way to sort it by going through all merge commits (as we
only publish images for them) of the current git listing those that are
present in the quay repo.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>