runtime: implement initdata setting and delivering
This commit adds the implementation of initdata setting on the runtime
side. If initdata annotation is given, the kata-runtime will ensure that
the first RPC called to kata-agent is the SetInitdata. This commit also
implements both TDX and SNP platform initdata setting.

To align with both legacy kata and the CoCo scenarios where initdata is
not used, an empty SetInitdata call will be performed if no cc_init_data
annotation is set.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>