agent: add initdata implementation
This commit adds the handling of initdata on the kata-agent side. Also
changes the launch logic of guest-components.

As we assume that the first api called by kata-runtime is SetInitData,
thus the guest-components would be launched inside the handler of
SetInitData. On the runtime side, if no initdata annotation is set, an
empty SetInitData request will be performed, thus guest components would
be launch with default config.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>