runtime: don't use machtype default for remote-hyp
The host's default machine type doesn't make sense for remote
hypervisor, since the remote component defines the machines. In Cloud
API Adaptor this is causing problems, since it defines it's own default
which gets overridden by a "q35" machinetype annotation from the host.
We don't want to overwrite the CAA default with a runtime default value.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>