runtime: Ensure Confidential Guests use shared_fs=none
While we do have the option to, maybe, use virtio-9p for Confidential
Guests, that basically means sharing content between host and guest.

Instead of allowing users to shoot themselves in the foot, let's enforce
shared_fs=none is used for Confidential Guests, and it also brings in a
little bit of extra security that if containerd and Kata Containers get
maliciously misconfigured by an administrator we at least don't even
start the Confidential Guest.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>