ci: Enable starting kata agent as local process
The existing setup tries to start kata agent as local process.
Since a lot of dependent components which are part of the UVM rootfs
are not used in this setup, this commit works around the issues:
a. Install a local copy of a relaxed 'allow-all.rego' to initialize
the agent policy
b. Remove coco attestaion guest components from the installed
kata artifacts so that agent does not try to start these processes
c. Install bats

This setup tries to start kata-agent in near vanilla mode with agent
policy.

Fixes #10286

Signed-off-by: Sumedh Alok Sharma <sumsharma@microsoft.com>