build: Fix time dependency compilation error
This is one of the 3 PRs that enable kata-containers to be compiled by
the latest stable Rust (v1.80.1)

Kata is unable to be built with Rust v1.80.1 due, among other things,
to a compilation error with an indirect dependency.
This patch bumps it, that's why it's only changes on Cargo.locks.
Time is not a direct dependency, so no changes to Cargo.tomls.

The patch was achieved through cargo update -p time for all time >= 0.3.0 for all crates under Kata.

Fixes https://github.com/kata-containers/kata-containers/issues/10215

Signed-off-by: Emanuel Lima <emlima@redhat.com>