ci: Add the sandbox api testcases
A test case is added based on the intergrated cri-containerd case.
The difference between cri containerd integrated testcase and sandbox
api testcase is the "sandboxer" setting in the sandbox runtime handler.

If the "sandboxer" is set to "" or "podsandbox", then containerd will
use the legacy shimv2 api, and if the "sandboxer" is set to "shim", then
it will use the sandbox api to launch the pod.

In addition, add a containerd RC version. Because containerd officially
supports the sandbox api from version 2.0, and containerd 2.0 has not
yet been officially released, use the RC version as a test case first.

Signed-off-by: Fupan Li <fupan.lfp@antgroup.com>