build: Install tomli with pip in the docker image used for qemu build.
Python package tomli was added as a dependency in the qemu helper script for
setting up python virtual environment for building qemu.
Without this, the `mkvenv` command fails with:

```
mkvenv: Creating non-isolated virtual environment at 'pyvenv'

*** Ouch! ***

found no usable tomli, please install it
```

Add python3-tomli to the list of packages installed.
Also, upgrade base image for the Dockerfile to Ubuntu 22.04.
This is because qemu 9.1.0 has added a dependency for glib>=2.66.0
and Ubuntu 20.04 comes with glib 2.64.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>