workflows: Build components with MEASURED_ROOTFS=yes
This is the second step in order to get measured rootfs back in place.

This used to be set in the good and old `CCv0` branch, but when we
did the `merge-to-main` effort it ended up not being brought in.

In the `CCv0` branch this was done only for `x86_64` architectures, but
it doesn't hurt to be done for all the supported architectures which are
using a rootfs-image (so, no ppc64le support for this yet), as done as
part of this commit.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>