workflows: Temporarily stop building OVMF (revert me ASAP)
OVMF / OVMF-SEV are failing to build due to a dependency that was made
private.  Those 2 packages are only used by the SNP (uses OVMF) and SEV
(uses OVMF SEV) respectively, and should be taken care by the respective
CI owners.

For now, in order to avoid blocking the upstream development, let's just
not build those and we can revert this PR as soon as the OVMF situation
is solved.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>