dragonball: Setup workspace and centralize RustVMM
Setup workspace in dragonball, move `dbs` crates one level up to be
managed as members of that workspace.

And centralize all RustVMM crates to `workspace.dependencies` to prevent
having multiple versions of each RustVMM crate, which is error-prone and
inconsistent. With this setup, updates on RustVMM crates would be much
easier.

Fixes: #10525

Signed-off-by: Ruoqing He <heruoqing@iscas.ac.cn>